feat(product): use `MockCollectionMetadata` in factory
