+++ libs/product/testing/src/factories/public_api.ts
@@ -2,7 +2,7 @@ export * from './product-image.factory';
 export * from './extension.factory';
 export * from './kind.factory';
 export * from './default-product.factory';
-export * from './collection-metadata.factory';
+export * from './collection/metadata.factory';
 export * from './filters/public_api';
 
 export { DaffDefaultProductFactory as DaffProductFactory } from './default-product.factory';
